feat: add tile rendering system
This commit got out of hand. One refactoring lead to another related but
not exactly the same, and it escalated. In the end there were so many
changes so splitting it all up into semantic commits would take far too
much time.

This introduces an area rendering and movement system. There are also
improvements to the projects files and in general a shit ton of
refactoring.